feat: origin 뷰 화면분할
